requirements.txt
GitPython
openai>=1.3.5
python-dotenv
